Separate Store from Conn
